Random

+++ view/src/application/ErrorHandling.java
@@ -5,13 +5,13 @@ public class ErrorHandling extends Thread{
   public ErrorHandling() {
   }
 	
-	public static String OutOfRangeX(int value) {
+	public static String OutOfRangeX(double value) {
 		Fraeskopf._setPositionX(0);
 		Fraeskopf._setPositionY(0);
 		return value + "Dieser Wert ist horizontal ausserhalb der Arbeitsflaeche, geben Sie einen Wert unter 1400 ein.";
 	}
 	
-	public static String OutOfRangeY(int value) {
+	public static String OutOfRangeY(double value) {
 		Fraeskopf._setPositionY(0);
 		Fraeskopf._setPositionX(0);
 		return value + "Dieser Wert ist horizontal ausserhalb der Arbeitsflaeche, geben Sie einen Wert unter 1050 ein.";
